requirements.txt
h5py
wandb
pillow
lightning
hydra-core
torchmetrics
